requirements.txt
Flask==3.0.0
cx_Oracle==8.3.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2

